ApplyMapGridState moved from map manager to the MapGridComponent.
