+++ Robust.Shared/GameStates/GameStateMapData.cs
@@ -19,8 +19,11 @@ namespace Robust.Shared.GameStates
         [Serializable, NetSerializable]
         public struct GridDatum
         {
+            // TransformComponent State
             public readonly MapCoordinates Coordinates;
             public readonly Angle Angle;
+
+            // MapGridComponent State
             public readonly ChunkDatum[] ChunkData;
 
             public GridDatum(ChunkDatum[] chunkData, MapCoordinates coordinates, Angle angle)
